[Packages] Package list rewrite (#1647)
* [Packages] Name unity

* [Packages] Starting over

* [Database] Collate nocase when getting packages

* [Packages] Load packages into list

* [Packages] Add search controller to list

* [Packages] Actually set _isInstalled when we look it up

* [UI] Add half-height presentation controller

* [UI] Remove shadow from half height presentation, was causing issues

* [UI] Bold header view's action buton is hidden by default

* [Packages] Demo of what filters will look like

* [UI] Make half presentation controller work with different heights

* [Packages] Don't show filter count on sort section

* [UI] Change corner radius back, it was hurting my eyes

* [Database] Don't bother ordering if we're going to do it again later

* [Packages] Report changes in filter to package list controller

* [Packages] Filter packages at the end of packagesFromSource with a ZBPackageFIlter

* [Packages] Sort after filtering

* [Database] Remove role restrictions from database

* [Packages] Set sort order in filter pane

* [Packages] Show checkmark on selected sort order

* [Packages] Load packages async

* [Packages] Add name as a secondary sort descriptor to others

* [Packages] "Size" sorts by download size on remote repositories

* [Packages] Don't show sections filter if the user cannot set sections

* [Packages] Update the UI slightly

* [Packages] Add a swipe gesture to partial presentation controllers

* [Packages] Change close button for filter controller

* [UI] Make a common selection controller for future use

Will convert other selection controllers in the future

* [Packages] Apply section filters

* [UI] Fix normal selection style

* [Packages]

* [Packages] Nicer reload animation

Don't tell andrew!

* [Packages] Change date string and sorting for local source

* [Packages] Filter by search term

* [Database] Fix misaligned columns

* [Packages] Move installedDate to basePackage so we can sort without fetching

* [Packages] Move install/download size into base package

* [Packages] Easier package size conversion

* [Packages] Make the filter button filled when filter is active

* [UI] Hide excess table view separators

* [Settings] Add role selector

* [Packages] Save fitlers to preferences

* [Packages] Add updates section

* [UI] Fix packages spinner theming

* [Database] Be sure to match source in the inner join so that no duplicates show up

* [Packages] Re-add swipe actions

* [Packages] Upadte installed packages list as we're importing the local source

* [Packages] Get updates working properly

* [Database] Sort updates when returning them

* [Packages] Change how the cells work a bit

* [Packages] Make the upgrade all button work

* [Sources] Set refreshInProgress NO when finishing source refresh

* [Sources] Replace base sources with their full counterparts when finished loading

* [Packages] Show correct package when selecting a package from updates

* [Database] Keep track of uncategorized packages in section readout

* [Sources] Add all packages back to section readout

* [Sources] Fix for cells not lining up

* [Packages] Make version selectors use strings instead of package instances

* [Packages] Import packages if there is no import/modified date

* [Packages] Sort lesser and greaterVersions

* [Database] Make isPackageAvailable work

* [Database] Add some missing statement resets

* [Meta] Cleanup

* [Downloads] Only use caching if package count is > 0

* [Pacakages] Add border to icon view in cells and change corner radius

* [Packages] If section is uncategorized, set it as such

* [Sources] Put automatic source refreshes on another queue

* [Database] Don't check updates for packages we know aren't going to have any

+++ Zebra/Managers/Delegates/ZBSourceDelegate.h
@@ -24,7 +24,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)finishedDownloadForSource:(ZBBaseSource *)source;
 - (void)startedImportForSource:(ZBBaseSource *)source;
 - (void)finishedImportForSource:(ZBBaseSource *)source;
-- (void)updatesAvailable:(int)numberOfUpdates;
+- (void)updatesAvailable:(NSUInteger)numberOfUpdates;
 - (void)finishedSourceRefresh;
 - (void)addedSources:(NSSet <ZBBaseSource *> *)sources;
 - (void)removedSources:(NSSet <ZBBaseSource *> *)sources;
